+++ CHANGES.rst
@@ -5,7 +5,7 @@ Change log for sw.allotmentclub
 5.11.2 (unreleased)
 ===================
 
-- Nothing changed yet.
+- Switch Arbeitsstunden from 10 € to 25 €.
 
 
 5.11.1 (2023-09-18)